requirements.txt
aiogram==3.*
python-dotenv
aiohttp_socks
